// src/gql/graphql.ts
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
// ...
export type Scalars = {
  ID: { input: string; output: string; }
  String: { input: string; output: string; }
  Boolean: { input: boolean; output: boolean; }
  Int: { input: number; output: number; }
  Float: { input: number; output: number; }
};

export type AffectedRowsOutput = {
  __typename?: 'AffectedRowsOutput';
  count: Scalars['Int']['output'];
};

export type AggregateAttestation = {
  __typename?: 'AggregateAttestation';
  _avg?: Maybe<AttestationAvgAggregate>;
  _count?: Maybe<AttestationCountAggregate>;
  _max?: Maybe<AttestationMaxAggregate>;
  _min?: Maybe<AttestationMinAggregate>;
  _sum?: Maybe<AttestationSumAggregate>;
};
// ...
export type Attestation = {
  __typename?: 'Attestation';
  attester: Scalars['String']['output'];
  data: Scalars['String']['output'];
  decodedDataJson: Scalars['String']['output'];
  expirationTime: Scalars['Int']['output'];
  id: Scalars['String']['output'];
  ipfsHash: Scalars['String']['output'];
  isOffchain: Scalars['Boolean']['output'];
  recipient: Scalars['String']['output'];
  refUID: Scalars['String']['output'];
  revocable: Scalars['Boolean']['output'];
  revocationTime: Scalars['Int']['output'];
  revoked: Scalars['Boolean']['output'];
  schema: Schema;
  schemaId: Scalars['String']['output'];
  time: Scalars['Int']['output'];
  timeCreated: Scalars['Int']['output'];
  txid: Scalars['String']['output'];
};

export type AttestationAvgAggregate = {
  __typename?: 'AttestationAvgAggregate';
  expirationTime?: Maybe<Scalars['Float']['output']>;
  revocationTime?: Maybe<Scalars['Float']['output']>;
  time?: Maybe<Scalars['Float']['output']>;
  timeCreated?: Maybe<Scalars['Float']['output']>;
};

export type AttestationAvgOrderByAggregateInput = {
  expirationTime?: InputMaybe<SortOrder>;
  revocationTime?: InputMaybe<SortOrder>;
  time?: InputMaybe<SortOrder>;
  timeCreated?: InputMaybe<SortOrder>;
};

export type AttestationCountAggregate = {
  __typename?: 'AttestationCountAggregate';
  _all: Scalars['Int']['output'];
  attester: Scalars['Int']['output'];
  data: Scalars['Int']['output'];
  decodedDataJson: Scalars['Int']['output'];
  expirationTime: Scalars['Int']['output'];
  id: Scalars['Int']['output'];
  ipfsHash: Scalars['Int']['output'];
  isOffchain: Scalars['Int']['output'];
  recipient: Scalars['Int']['output'];
  refUID: Scalars['Int']['output'];
  revocable: Scalars['Int']['output'];
  revocationTime: Scalars['Int']['output'];
  revoked: Scalars['Int']['output'];
  schemaId: Scalars['Int']['output'];
  time: Scalars['Int']['output'];
  timeCreated: Scalars['Int']['output'];
  txid: Scalars['Int']['output'];
};
// ...
export type AttestationSumAggregate = {
  __typename?: 'AttestationSumAggregate';
  expirationTime?: Maybe<Scalars['Int']['output']>;
  revocationTime?: Maybe<Scalars['Int']['output']>;
  time?: Maybe<Scalars['Int']['output']>;
  timeCreated?: Maybe<Scalars['Int']['output']>;
};
// ...
export type AttestationWhereInput = {
  AND?: InputMaybe<Array<AttestationWhereInput>>;
  NOT?: InputMaybe<Array<AttestationWhereInput>>;
  OR?: InputMaybe<Array<AttestationWhereInput>>;
  attester?: InputMaybe<StringFilter>;
  data?: InputMaybe<StringFilter>;
  decodedDataJson?: InputMaybe<StringFilter>;
  expirationTime?: InputMaybe<IntFilter>;
  id?: InputMaybe<StringFilter>;
  ipfsHash?: InputMaybe<StringFilter>;
  isOffchain?: InputMaybe<BoolFilter>;
  recipient?: InputMaybe<StringFilter>;
  refUID?: InputMaybe<StringFilter>;
  revocable?: InputMaybe<BoolFilter>;
  revocationTime?: InputMaybe<IntFilter>;
  revoked?: InputMaybe<BoolFilter>;
  schema?: InputMaybe<SchemaRelationFilter>;
  schemaId?: InputMaybe<StringFilter>;
  time?: InputMaybe<IntFilter>;
  timeCreated?: InputMaybe<IntFilter>;
  txid?: InputMaybe<StringFilter>;
};

export type AttestationWhereUniqueInput = {
  id?: InputMaybe<Scalars['String']['input']>;
};

export type BoolFieldUpdateOperationsInput = {
  set?: InputMaybe<Scalars['Boolean']['input']>;
};

export type BoolFilter = {
  equals?: InputMaybe<Scalars['Boolean']['input']>;
  not?: InputMaybe<NestedBoolFilter>;
};
// ...
export type IntFilter = {
  equals?: InputMaybe<Scalars['Int']['input']>;
  gt?: InputMaybe<Scalars['Int']['input']>;
  gte?: InputMaybe<Scalars['Int']['input']>;
  in?: InputMaybe<Array<Scalars['Int']['input']>>;
  lt?: InputMaybe<Scalars['Int']['input']>;
  lte?: InputMaybe<Scalars['Int']['input']>;
  not?: InputMaybe<NestedIntFilter>;
  notIn?: InputMaybe<Array<Scalars['Int']['input']>>;
};
// ...
export type Schema = {
  __typename?: 'Schema';
  _count?: Maybe<SchemaCount>;
  attestations: Array<Attestation>;
  creator: Scalars['String']['output'];
  id: Scalars['String']['output'];
  index: Scalars['String']['output'];
  resolver: Scalars['String']['output'];
  revocable: Scalars['Boolean']['output'];
  schema: Scalars['String']['output'];
  schemaNames: Array<SchemaName>;
  time: Scalars['Int']['output'];
  txid: Scalars['String']['output'];
};
// ...
export type SchemaRelationFilter = {
  is?: InputMaybe<SchemaWhereInput>;
  isNot?: InputMaybe<SchemaWhereInput>;
};
// ...
export type StringFilter = {
  contains?: InputMaybe<Scalars['String']['input']>;
  endsWith?: InputMaybe<Scalars['String']['input']>;
  equals?: InputMaybe<Scalars['String']['input']>;
  gt?: InputMaybe<Scalars['String']['input']>;
  gte?: InputMaybe<Scalars['String']['input']>;
  in?: InputMaybe<Array<Scalars['String']['input']>>;
  lt?: InputMaybe<Scalars['String']['input']>;
  lte?: InputMaybe<Scalars['String']['input']>;
  mode?: InputMaybe<QueryMode>;
  not?: InputMaybe<NestedStringFilter>;
  notIn?: InputMaybe<Array<Scalars['String']['input']>>;
  startsWith?: InputMaybe<Scalars['String']['input']>;
};
// ...
export type AttestationWithCoreFieldsFragment = { __typename?: 'Attestation', id: string, time: number, attester: string, recipient: string, decodedDataJson: string } & { ' $fragmentName'?: 'AttestationWithCoreFieldsFragment' };

export type AttestationsQueryVariables = Exact<{
  where?: InputMaybe<AttestationWhereInput>;
  take?: InputMaybe<Scalars['Int']['input']>;
  skip?: InputMaybe<Scalars['Int']['input']>;
}>;


export type AttestationsQuery = { __typename?: 'Query', attestations: Array<(
    { __typename?: 'Attestation' }
    & { ' $fragmentRefs'?: { 'AttestationWithCoreFieldsFragment': AttestationWithCoreFieldsFragment } }
  )> };

export type AggregateAttestationQueryVariables = Exact<{
  where?: InputMaybe<AttestationWhereInput>;
}>;


export type AggregateAttestationQuery = { __typename?: 'Query', aggregateAttestation: { __typename?: 'AggregateAttestation', _count?: { __typename?: 'AttestationCountAggregate', id: number } | null } };

export type RecipientAttestationsQueryVariables = Exact<{
  where?: InputMaybe<AttestationWhereInput>;
}>;


export type RecipientAttestationsQuery = { __typename?: 'Query', attestations: Array<(
    { __typename?: 'Attestation' }
    & { ' $fragmentRefs'?: { 'AttestationWithCoreFieldsFragment': AttestationWithCoreFieldsFragment } }
  )> };

export const AttestationWithCoreFieldsFragmentDoc = {"kind":"Document","definitions":[{"kind":"FragmentDefinition","name":{"kind":"Name","value":"AttestationWithCoreFields"},"typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"Attestation"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","name":{"kind":"Name","value":"time"}},{"kind":"Field","name":{"kind":"Name","value":"attester"}},{"kind":"Field","name":{"kind":"Name","value":"recipient"}},{"kind":"Field","name":{"kind":"Name","value":"decodedDataJson"}}]}}]} as unknown as DocumentNode<AttestationWithCoreFieldsFragment, unknown>;
export const AttestationsDocument = {"kind":"Document","definitions":[{"kind":"OperationDefinition","operation":"query","name":{"kind":"Name","value":"Attestations"},"variableDefinitions":[{"kind":"VariableDefinition","variable":{"kind":"Variable","name":{"kind":"Name","value":"where"}},"type":{"kind":"NamedType","name":{"kind":"Name","value":"AttestationWhereInput"}}},{"kind":"VariableDefinition","variable":{"kind":"Variable","name":{"kind":"Name","value":"take"}},"type":{"kind":"NamedType","name":{"kind":"Name","value":"Int"}}},{"kind":"VariableDefinition","variable":{"kind":"Variable","name":{"kind":"Name","value":"skip"}},"type":{"kind":"NamedType","name":{"kind":"Name","value":"Int"}}}],"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"attestations"},"arguments":[{"kind":"Argument","name":{"kind":"Name","value":"orderBy"},"value":{"kind":"ObjectValue","fields":[{"kind":"ObjectField","name":{"kind":"Name","value":"time"},"value":{"kind":"EnumValue","value":"desc"}}]}},{"kind":"Argument","name":{"kind":"Name","value":"where"},"value":{"kind":"Variable","name":{"kind":"Name","value":"where"}}},{"kind":"Argument","name":{"kind":"Name","value":"take"},"value":{"kind":"Variable","name":{"kind":"Name","value":"take"}}},{"kind":"Argument","name":{"kind":"Name","value":"skip"},"value":{"kind":"Variable","name":{"kind":"Name","value":"skip"}}}],"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"FragmentSpread","name":{"kind":"Name","value":"AttestationWithCoreFields"}}]}}]}},{"kind":"FragmentDefinition","name":{"kind":"Name","value":"AttestationWithCoreFields"},"typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"Attestation"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","name":{"kind":"Name","value":"time"}},{"kind":"Field","name":{"kind":"Name","value":"attester"}},{"kind":"Field","name":{"kind":"Name","value":"recipient"}},{"kind":"Field","name":{"kind":"Name","value":"decodedDataJson"}}]}}]} as unknown as DocumentNode<AttestationsQuery, AttestationsQueryVariables>;
export const AggregateAttestationDocument = {"kind":"Document","definitions":[{"kind":"OperationDefinition","operation":"query","name":{"kind":"Name","value":"AggregateAttestation"},"variableDefinitions":[{"kind":"VariableDefinition","variable":{"kind":"Variable","name":{"kind":"Name","value":"where"}},"type":{"kind":"NamedType","name":{"kind":"Name","value":"AttestationWhereInput"}}}],"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"aggregateAttestation"},"arguments":[{"kind":"Argument","name":{"kind":"Name","value":"where"},"value":{"kind":"Variable","name":{"kind":"Name","value":"where"}}}],"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"_count"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}}]}}]}}]}}]} as unknown as DocumentNode<AggregateAttestationQuery, AggregateAttestationQueryVariables>;
export const RecipientAttestationsDocument = {"kind":"Document","definitions":[{"kind":"OperationDefinition","operation":"query","name":{"kind":"Name","value":"RecipientAttestations"},"variableDefinitions":[{"kind":"VariableDefinition","variable":{"kind":"Variable","name":{"kind":"Name","value":"where"}},"type":{"kind":"NamedType","name":{"kind":"Name","value":"AttestationWhereInput"}}}],"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"attestations"},"arguments":[{"kind":"Argument","name":{"kind":"Name","value":"orderBy"},"value":{"kind":"ObjectValue","fields":[{"kind":"ObjectField","name":{"kind":"Name","value":"time"},"value":{"kind":"EnumValue","value":"desc"}}]}},{"kind":"Argument","name":{"kind":"Name","value":"where"},"value":{"kind":"Variable","name":{"kind":"Name","value":"where"}}}],"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"FragmentSpread","name":{"kind":"Name","value":"AttestationWithCoreFields"}}]}}]}},{"kind":"FragmentDefinition","name":{"kind":"Name","value":"AttestationWithCoreFields"},"typeCondition":{"kind":"NamedType","name":{"kind":"Name","value":"Attestation"}},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","name":{"kind":"Name","value":"time"}},{"kind":"Field","name":{"kind":"Name","value":"attester"}},{"kind":"Field","name":{"kind":"Name","value":"recipient"}},{"kind":"Field","name":{"kind":"Name","value":"decodedDataJson"}}]}}]} as unknown as DocumentNode<RecipientAttestationsQuery, RecipientAttestationsQueryVariables>;